--- packages/material/src/hooks.js
@@ -1,7 +1,0 @@
-export function init(config) {
-    if (config) {
-        Object.assign(this.config, config);
-    }
-}
-
-export function start() {}
