+++ packages/material/src/init.js
@@ -1,0 +1,1 @@
+export function init() {}
