+++ packages/react/src/init.expo.js
@@ -1,0 +1,4 @@
+import 'expo/build/Expo.fx';
+import { init, start, unmount } from './init.native.js';
+
+export { init, start, unmount };
